replace long func with constant
